Version bump to 0.1.0 because I guess it is at a
stable enough state to be called "stable". The API functionality has
been fully covered and should be available to use. However it literally
has no users yet so I will not label it as "1.0" because that would make
it look like it has achieved reliability, but as nobody has tested it in
practice, there is no evidence of that as of now. Besides, even if the
library was proved to be stable, it is not at the point where I have
originally planned to develop it, which is the version which I would
personally call "1.0". But ya here's a version bump.